requirements.txt
streamlit
pandas
gspread
google-auth
requests
Pillow
numpy
fpdf2
qrcode
streamlit-drawable-canvas
